requirements.txt
aistudio-sdk==0.3.8
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
bce-python-sdk==0.9.68
bidict==0.23.1
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
chardet==7.4.1
charset-normalizer==3.4.7
click==8.3.1
clr_loader==0.2.10
colorama==0.4.6
colorlog==6.10.1
crc32c==2.8
dnspython==2.8.0
filelock==3.25.2
Flask==3.1.3
Flask-SocketIO==5.6.1
fsspec==2026.3.0
future==1.0.0
greenlet==3.3.2
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.9.2
idna==3.11
imagesize==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
modelscope==1.35.3
mpmath==1.3.0
mss==10.1.0
networkx==3.6.1
numpy==2.4.4
opencv-contrib-python==4.10.0.84
opencv-python==4.13.0.92
opt-einsum==3.3.0
packaging==26.0
pandas==3.0.2
pillow==12.1.1
prettytable==3.17.0
protobuf==7.34.1
proxy_tools==0.1.0
psutil==7.2.2
py-cpuinfo==9.0.0
pyclipper==1.4.0
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
PyDirectInput==1.0.4
Pygments==2.20.0
PyMsgBox==2.0.1
pypdfium2==5.6.0
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
pytesseract==0.3.13
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-engineio==4.13.1
python-socketio==5.16.1
pytweening==1.2.0
pywin32==311
PyYAML==6.0.2
requests==2.33.1
rich==14.3.3
ruamel.yaml==0.19.1
safetensors==0.7.0
shapely==2.1.2
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
sympy==1.14.0
torch==2.11.0
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
ujson==5.12.0
urllib3==2.6.3
wcwidth==0.6.0
Werkzeug==3.1.7
wsproto==1.3.2
